LayoutTests/inspector/debugger/break-on-uncaught-exception-throw-in-promise-then-with-catch-expected.txt

CONSOLE MESSAGE: line 66: testThrowingInPromiseThenWithCatch
CONSOLE MESSAGE: line 68: in promise
CONSOLE MESSAGE: line 71: in promise.then
CONSOLE MESSAGE: line 7: throwing TestError
CONSOLE MESSAGE: line 74: in promise.catch
CONSOLE MESSAGE: line 75: DONE
Checking pause locations when pausing on uncaught exceptions.

PAUSE #1 AT: doThrow:8:22
